增加了课程删除与分类删除的限制条件

+++ service/service_edu/src/main/java/com/mqd/eduservice/service/impl/EduSubjectServiceImpl.java
@@ -1,16 +1,20 @@
 package com.mqd.eduservice.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.mqd.eduservice.mapper.EduCourseMapper;
+import com.mqd.eduservice.pojo.EduCourse;
 import com.mqd.eduservice.pojo.EduSubject;
 import com.mqd.eduservice.mapper.EduSubjectMapper;
 import com.mqd.eduservice.pojo.dto.SubjectDto;
 import com.mqd.eduservice.service.EduSubjectService;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.mqd.exception.CustomException;
+import com.sun.xml.internal.messaging.saaj.packaging.mime.util.QEncoderStream;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.annotation.Resource;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
@@ -27,6 +31,9 @@ import java.util.List;
 @Slf4j
 public class EduSubjectServiceImpl extends ServiceImpl<EduSubjectMapper, EduSubject> implements EduSubjectService {
 
+    @Resource
+    private EduCourseMapper courseMapper;
+
     @Override
     public List<SubjectDto> getSubjectTree() {
         List<SubjectDto> list = new ArrayList<>();
@@ -39,13 +46,11 @@ public class EduSubjectServiceImpl extends ServiceImpl<EduSubjectMapper, EduSubj
                 list.add(subjectDto);
             }
         });
-        res.forEach(n->{
-            list.forEach(m->{
-                if (n.getParentId().equals(m.getId())) {
-                    m.add(n);
-                }
-            });
-        });
+        res.forEach(n-> list.forEach(m->{
+            if (n.getParentId().equals(m.getId())) {
+                m.add(n);
+            }
+        }));
         return list;
     }
 
@@ -54,16 +59,29 @@ public class EduSubjectServiceImpl extends ServiceImpl<EduSubjectMapper, EduSubj
 
         EduSubject eduSubject = baseMapper.selectById(id);
         if (!eduSubject.getParentId().equals("0")) {
-            //当其是次级目录的时候直接删除
-             baseMapper.deleteById(id);
+            // 先检查该分类下有没有课程存在，如果有则不能删除
+            //当其是次级目录且分类下没有课程存在的时候直接删除
+            QueryWrapper<EduCourse> wrapper = new QueryWrapper<>();
+            wrapper.eq("subject_id",id);
+            Integer integer = courseMapper.selectCount(wrapper);
+            if (integer > 0){
+                throw new CustomException("当前该分类下存在课程,无法删除!");
+            }
+            baseMapper.deleteById(id);
             log.info(eduSubject.getId()+":"+eduSubject.getTitle()+"--->已经成功删除");
         }else {
             //当其不是次级目录时，先判断是否存在子目录
-            QueryWrapper<EduSubject> wrapper = new QueryWrapper<EduSubject>();
+            QueryWrapper<EduSubject> wrapper = new QueryWrapper<>();
             wrapper.eq("parent_id", id);
             Integer integer = baseMapper.selectCount(wrapper);
             if (integer>0){
                 throw new CustomException("当前存在子分类,无法删除!");
+            }
+            QueryWrapper<EduCourse> courseWrapper = new QueryWrapper<>();
+            wrapper.eq("subject_parent_id",id);
+            Integer courseCount = courseMapper.selectCount(courseWrapper);
+            if (courseCount > 0){
+                throw new CustomException("当前该分类下存在课程,无法删除!");
             }
             int i = baseMapper.deleteById(id);
             return i > 0;
